feat(codeShare): implement code sharing module with UI and service enhancements
- Add new CodeShare module with REST APIs for code management.
- Integrate with Nacos for configuration management and service discovery.
- Implement pagination and filtering for code list retrieval.
- Secure code sharing with user authentication and authorization.
- Improve code quality with MyBatis Plus optimizations and custom handlers.
- Add unit tests for CodeShare service and controller.
- Refactor existing code for better maintainability and performance.

BREAKING CHANGE: Use Long instead of Integer for page and size parameters in PageResult.

+++ muyan-common/muyan-common-core/src/main/java/com/muyan/domain/PageResult.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 /**
  * 分页查询结果
+ *
  * @author huliua
  * @version 1.0
  * @date 2024-05-05 17:00
@@ -16,17 +17,17 @@ public class PageResult<T> {
     /**
      * 总记录数
      */
-    private Integer total;
+    private Long total;
 
     /**
      * 总分页数
      */
-    private Integer pageNum;
+    private Long pageNum;
 
     /**
      * 每页记录数
      */
-    private Integer pageSize;
+    private Long pageSize;
 
     /**
      * 分页查询结果
